Implemented language flags
Each channel/stream now has language flag indicators.
There can be two different flags for the channel language and the
channel broadcaster language. If both share the same value, the latter
will be hidden.

Also slightly changed the style of the featured channels and
changed the title attribute of the StreamItemComponent to be only
visible on the preview image.