The RESTAdapter now supports dynamic URLs
Ember-Data is still lacking this feature, so it had to be done in this
kind of improvised way.

This change enables the implementation of the channel following and
channel subscription-check mechanics, as well as the listing of channel
videos and teams the channel belongs to. These kind of requests require
the URL of the twitch.tv API endpoint to contain the user/channel name.