Temporarily fix stream reference bugs
This is just a workaround to prevent stream records from being
mis-referenced. Now the issue is, that those records will never be
refreshed when visiting these "fixed" routes. Ideally, the stream
records should be refreshed by calling ".refresh()", which then triggers
a fetch in the data store, but this can't be done right now.

TODO:
- Change the stream serializer's primary key to "name" (see channels)
- Rewrite buildURL method of the twitchAdapter
- Fix streamFollowed, searchStream, searchGame models/serializers