Refactor subclasses of twitchStream and twitchGame
This fixes the issue described in ecd8918.

twitchStreamsFollowed, twitchSearchStream and twitchSearchGame are
no longer subclasses of twitchStream or twitchGame and instead act like
twitchStreamsFeatured: they now contain a belongsTo relationship to the
actual/intended record.
All stream or game queries to the store will now create/update records
of the correct type and will no longer use subclass-types. This also
means that we're able to .refresh() twitchStreams correctly even if the
record was created by a twitchStreamsFollowed query.