Immediately refresh stream record on stream start
Sometimes the twitch API returns a user record instead of a channel
record embedded into a stream record. This seems to be only happening
while querying for multiple streams. A stream then doesn't have any
viewer, views or game informations.
This change should fix the problem by reloading the stream record as
soon as the user is starting a stream, so the livestreamer popup can
show informative data.