Fix livestreamer validation. Fix #121
Sometimes, the exit event of the child process gets executed first, so
the validation promise gets rejected. Ignore exit code `0` to fix this.

Also implement a StreamOutputBuffer, in case the output is being
returned in chunks. The StreamOutputBuffer buffers each chunk and splits
everything into lines, so each line can be parsed as a whole.

The StreamOutputBuffer will also be used by future IRC implemenations.